requirements.txt
Django==1.7
Pillow==2.3.0
dj-database-url==0.3.0
django-taggit==0.11.2
gunicorn==18.0
psycopg2==2.5.2
whitenoise==1.0.6
wsgiref==0.1.2
